Change user schema to support admin and handle breaking changes
Breaking changes occur in the signup action creator, router, and test files. The cause of breaking changes is the admin boolean in the User model is now a required field. Therefore, the req.body must contain  admin: bool. Finally, do not allow admin users to sign in reg-user login. I will have to make changes to client-side auth, to render a different UI for adminUser.